feat: insere verificação de parametros da url no servidor

+++ gerenciamento-projetos/backend/src/controllers/projects.controller.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import { IProjectService } from "../services/projects.service";
+import { httpStatus } from "../utils/httpStatus";
 
 export class ProjectsController {
   private service: IProjectService;
@@ -16,6 +17,12 @@ export class ProjectsController {
   public async findOne(req: Request, res: Response) {
     const id = req.params["projectId"];
 
+    if (!id) {
+      return res.status(httpStatus.BadRequest).json({
+        error: "É necessário um id de projeto para completar a ação.",
+      });
+    }
+
     const { status, message } = await this.service.findOne({ id });
     res.status(status).json(message);
   }
@@ -28,16 +35,29 @@ export class ProjectsController {
   public async update(req: Request, res: Response) {
     const projectId = req.params["projectId"];
 
+    if (!projectId) {
+      return res.status(httpStatus.BadRequest).json({
+        error: "É necessário um id de projeto para completar a ação.",
+      });
+    }
+
     const { status, message } = await this.service.update(projectId, req.body);
     res.status(status).json(message);
   }
 
   public async remove(req: Request, res: Response) {
     const projectId = req.params["projectId"];
+
+    if (!projectId) {
+      return res.status(httpStatus.BadRequest).json({
+        error: "É necessário um id de projeto para completar a ação.",
+      });
+    }
+
     const { status, message } = await this.service.remove(projectId);
 
     if (message) {
-      res.status(status!).json(message);
+      return res.status(status!).json(message);
     }
 
     res.sendStatus(status!);
