+++ gerenciamento-projetos/backend/src/controllers/projectsUsers.controller.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import { IProjectsUsersService } from "../services/projetcsUsers.service";
+import { httpStatus } from "../utils/httpStatus";
 
 export class ProjectsUsersController {
   private service: IProjectsUsersService;
@@ -10,6 +11,13 @@ export class ProjectsUsersController {
 
   public async findUsersInProject(req: Request, res: Response) {
     const projectId = req.params["projectId"];
+
+    if (!projectId) {
+      return res.status(httpStatus.BadRequest).json({
+        error: "É necessário um id de projeto para completar a ação.",
+      });
+    }
+
     const { status, message } = await this.service.findUsersInProject(
       projectId
     );
@@ -18,6 +26,12 @@ export class ProjectsUsersController {
 
   public async insertUserInProject(req: Request, res: Response) {
     const projectId = req.params["projectId"];
+
+    if (!projectId) {
+      return res.status(httpStatus.BadRequest).json({
+        error: "É necessário um id de projeto para completar a ação.",
+      });
+    }
 
     const { status, message } = await this.service.insertUserInProject(
       projectId,
@@ -29,6 +43,18 @@ export class ProjectsUsersController {
   public async removeUserFromProject(req: Request, res: Response) {
     const projectId = req.params["projectId"];
     const userId = req.params["userId"];
+
+    if (!projectId) {
+      return res.status(httpStatus.BadRequest).json({
+        error: "É necessário um id de projeto para completar a ação.",
+      });
+    }
+
+    if (!userId) {
+      return res.status(httpStatus.BadRequest).json({
+        error: "É necessário um id de usuário para completar a ação.",
+      });
+    }
 
     const { status } = await this.service.removeUserFromProject(
       projectId,
